Update a bunch of vars
